backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
httpx
loguru
python-dotenv
pytest
tinydb
apscheduler
